+++ app/src/main/java/com/example/shopapp/presentation/home/composable/HomeTopBar.kt
@@ -25,7 +25,7 @@ fun HomeTopBar() {
         modifier = Modifier
             .fillMaxWidth()
             .background(MaterialTheme.colors.background)
-            .padding(vertical = 10.dp),
+            .padding(vertical = 15.dp),
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.SpaceBetween
     ) {
