+++ app/src/main/java/com/example/shopapp/presentation/navigation/composable/BottomBarNavigation.kt
@@ -1,0 +1,31 @@
+package com.example.shopapp.presentation.navigation.composable
+
+import androidx.compose.material.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.navigation.NavController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import com.example.shopapp.domain.model.BottomBarItem
+
+@Composable
+fun BottomBarNavigation(
+    items: List<BottomBarItem>,
+    navController: NavController,
+    onItemClick: (BottomBarItem) -> Unit
+) {
+    val backStackEntry = navController.currentBackStackEntryAsState()
+    BottomNavigation(
+        backgroundColor = MaterialTheme.colors.background
+    ) {
+        items.forEach { item ->
+            val selected = item.route == backStackEntry.value?.destination?.route
+            BottomNavigationItem(
+                selected = selected,
+                onClick = { onItemClick(item) },
+                selectedContentColor = Color.Black,
+                unselectedContentColor = Color.Gray,
+                icon = { BottomBarNavigationItem(item = item) }
+            )
+        }
+    }
+}
