+++ app/src/main/java/com/example/shopapp/presentation/navigation/composable/ShopAppNavigation.kt
@@ -1,10 +1,9 @@
-package com.example.shopapp.presentation.navigation
+package com.example.shopapp.presentation.navigation.composable
 
 import android.annotation.SuppressLint
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.rememberNavController
-import com.example.shopapp.presentation.common.composable.BottomBar
 
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
